Add resolving util for rAF and rIC which falls back to stubbing setTimeout when requestIdleCallback isn't supported
