+++ tests/unit/meta/Drag.ts
@@ -1,30 +1,14 @@
-import global from '@dojo/shim/global';
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
-import { stub, SinonStub } from 'sinon';
 import sendEvent from '../../support/sendEvent';
+import { createResolvers } from './../../support/util';
 import { v } from '../../../src/d';
 import { ProjectorMixin } from '../../../src/main';
 import Drag, { DragResults } from '../../../src/meta/Drag';
 import { WidgetBase } from '../../../src/WidgetBase';
 import { ThemeableMixin } from '../../../src/mixins/Themeable';
 
-let rAF: SinonStub;
-let rIC: SinonStub;
-
-function resolveRAF() {
-	for (let i = 0; i < rAF.callCount; i++) {
-		rAF.getCall(i).args[0]();
-	}
-	rAF.reset();
-}
-
-function resolveRIC() {
-	for (let i = 0; i < rIC.callCount; i++) {
-		rIC.getCall(i).args[0]();
-	}
-	rIC.reset();
-}
+const resolvers = createResolvers();
 
 const emptyResults: DragResults = {
 	delta: { x: 0, y: 0 },
@@ -35,13 +19,11 @@ registerSuite({
 	name: 'support/meta/Drag',
 
 	beforeEach() {
-		rAF = stub(global, 'requestAnimationFrame').returns(1);
-		rIC = stub(global, 'requestIdleCallback').returns(1);
+		resolvers.stub();
 	},
 
 	afterEach() {
-		rAF.restore();
-		rIC.restore();
+		resolvers.restore();
 	},
 
 	'standard rendering'() {
@@ -64,9 +46,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [ emptyResults, emptyResults ], 'should have been called twice, both empty results');
 
@@ -97,9 +78,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [ emptyResults, emptyResults ], 'should have been called twice, both empty results');
 
@@ -131,9 +111,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -151,8 +130,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -168,8 +146,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -185,8 +162,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -234,9 +210,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -254,8 +229,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -299,8 +273,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -316,8 +289,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -365,9 +337,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -385,8 +356,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -444,8 +414,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -489,9 +458,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -507,8 +475,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -524,8 +491,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -564,9 +530,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -584,8 +549,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -601,8 +565,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -618,8 +581,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -675,9 +637,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -695,8 +656,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -712,8 +672,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild!.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -729,8 +688,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -765,9 +723,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -785,8 +742,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -802,8 +758,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -819,8 +774,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -855,9 +809,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -875,8 +828,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -894,8 +846,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -911,8 +862,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -928,8 +878,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
@@ -972,9 +921,8 @@ registerSuite({
 		const widget = new TestWidget();
 		widget.append(div);
 
-		resolveRIC();
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerdown', {
 			eventInit: {
@@ -992,8 +940,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointermove', {
 			eventInit: {
@@ -1009,13 +956,11 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		widget.invalidate();
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		sendEvent(div.firstChild as Element, 'pointerup', {
 			eventInit: {
@@ -1031,8 +976,7 @@ registerSuite({
 			}
 		});
 
-		resolveRAF();
-		resolveRIC();
+		resolvers.resolve();
 
 		assert.deepEqual(dragResults, [
 			emptyResults,
